Update QoaDecoder to stream from std::io::Read
Instead of requiring all input bytes and all output bytes to be in
memory simultaneously, stream the input from a std::io::Read and
allow the output to be retrieved sample by sample. This reduces memory
consumption from O(2*N) to O(1) with regards to file size. Samples are
decoded one slice at a time and stored so the fixed memory overhead is
40 bytes per channel.

This is a large API breaking change. Since an IO error can occur while
decoding a new IoError variant is introduced and some existing variants
are replaced by the UnexpectedEoF error kind of io::Error.